Move jupyter command finder into interpreter dir
