+++ src/client/datascience/jupyter/interpreter/README.md
@@ -1,0 +1,1 @@
+# Contains code related to the interpreter(s) used to start Jupyter, get kernel specs, etc.
