+++ src/client/datascience/serviceRegistry.ts
@@ -31,8 +31,8 @@ import { NativeEditorProvider } from './interactive-ipynb/nativeEditorProvider';
 import { InteractiveWindow } from './interactive-window/interactiveWindow';
 import { InteractiveWindowCommandListener } from './interactive-window/interactiveWindowCommandListener';
 import { InteractiveWindowProvider } from './interactive-window/interactiveWindowProvider';
-import { JupyterCommandFactory } from './jupyter/jupyterCommand';
-import { JupyterCommandFinder } from './jupyter/jupyterCommandFinder';
+import { JupyterCommandFactory } from './jupyter/interpreter/jupyterCommand';
+import { JupyterCommandFinder } from './jupyter/interpreter/jupyterCommandFinder';
 import { JupyterDebugger } from './jupyter/jupyterDebugger';
 import { JupyterExecutionFactory } from './jupyter/jupyterExecutionFactory';
 import { JupyterExporter } from './jupyter/jupyterExporter';
