+++ src/test/datascience/execution.unit.test.ts
@@ -41,8 +41,8 @@ import { createDeferred } from '../../client/common/utils/async';
 import { Architecture } from '../../client/common/utils/platform';
 import { EXTENSION_ROOT_DIR } from '../../client/constants';
 import { Identifiers, PythonDaemonModule } from '../../client/datascience/constants';
-import { JupyterCommandFactory } from '../../client/datascience/jupyter/jupyterCommand';
-import { JupyterCommandFinder } from '../../client/datascience/jupyter/jupyterCommandFinder';
+import { JupyterCommandFactory } from '../../client/datascience/jupyter/interpreter/jupyterCommand';
+import { JupyterCommandFinder } from '../../client/datascience/jupyter/interpreter/jupyterCommandFinder';
 import { JupyterExecutionFactory } from '../../client/datascience/jupyter/jupyterExecutionFactory';
 import { KernelSelector } from '../../client/datascience/jupyter/kernels/kernelSelector';
 import { LiveKernelModel } from '../../client/datascience/jupyter/kernels/types';
